refactor: extract algorithm entry point into a separate file
